% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital-rates.R
\name{age_first_rep}
\alias{age_first_rep}
\title{Calculate median, minimum, and maximum age at first reproduction}
\usage{
age_first_rep(b)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}
}
\value{
Output is a tbl_df that gives the median, minimum, and maximum age at first reproduction for each study species in days and years. Sample size is also included.
}
\description{
Calculate median, minimum, and maximum age at first reproduction
}
\examples{
afr <- age_first_rep(lh)
}

