% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital-rates.R
\name{age_specific_fertility}
\alias{age_specific_fertility}
\title{Calculate age-specific fertility
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.}
\usage{
age_specific_fertility(b, f)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}

\item{f}{Name of the tbl_df containing fertility data generated by the function read_fert_table.}
}
\value{
Output is a tbl_df that includes the per capita fertility for each discrete age class for each study species. Also included are the number of animals and the "female-years" of observation on which the calculation is based.
}
\description{
Calculate age-specific fertility
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.
}
\examples{
asf <- age_specific_fertility(lh, fert)
}

