% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{cubroot_trans}
\alias{cubroot_trans}
\title{A cube-root transformation for scales, e.g., color gradients.}
\usage{
cubroot_trans()
}
\description{
A cube-root transformation for scales, e.g., color gradients.
}

