% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-errors.R
\name{find_bio_errors}
\alias{find_bio_errors}
\title{Find problems in Biography data.}
\usage{
find_bio_errors(x)
}
\arguments{
\item{x}{Name of the tbl_df containing biography data generated by the function read_bio_table.}
}
\description{
Returns a named list of tbl_df objects, including
\itemize{
\item error_dates: Rows with one or more bad dates
\item error_duplicates: Rows with duplicate animals
}
}
\examples{
bio_errors <- find_bio_errors(lh)
}

