% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_errors.R
\name{find_fert_errors}
\alias{find_fert_errors}
\title{Find problems in Fertility data.}
\usage{
find_fert_errors(x)
}
\arguments{
\item{x}{Name of the tbl_df containing fertility data generated by the function read_fert_table.}
}
\description{
Returns a named list of tbl_df objects, including
\itemize{
\item error_dates: Rows with one or more bad dates
}
}
\examples{
fert_errors <- find_fert_errors(fert)
}

