% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-errors.R
\name{find_first_born_errors}
\alias{find_first_born_errors}
\title{Find mothers with no corresponding Animal.Id entry.}
\usage{
find_first_born_errors(x)
}
\arguments{
\item{x}{Name of the tbl_df containing biography data generated by the function read_bio_table.}
}
\description{
Returns a named list Mom.Id records that don't match the Animal.Id records for each site
}
\examples{
first_born_errors <- find_first_born_errors(lh)
}

