% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_errors.R
\name{find_mom_errors}
\alias{find_mom_errors}
\title{Find mothers with no corresponding Animal.Id entry.}
\usage{
find_mom_errors(x)
}
\arguments{
\item{x}{Name of the tbl_df containing biography data generated by the function read_bio_table.}
}
\description{
Returns a named list Mom.Id records that don't match the Animal.Id records for each site
}
\examples{
mom_errors <- find_mom_errors(lh)
}

