% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital-rates.R
\name{get_age_class}
\alias{get_age_class}
\title{Calculates the age class of an animal based on its age and the median age at
first reproduction for the species.}
\usage{
get_age_class(age_years, mafr)
}
\arguments{
\item{age_years}{Age of the animal in years.}

\item{mafr}{Median age of first reproduction in years for the species.}
}
\value{
Output is a character string: "newborn", "juvenile", or "adult"
}
\description{
Calculates the age class of an animal based on its age and the median age at
first reproduction for the species.
}
\examples{
age_class <- get_age_class(5.1, 8.78589)
}

