% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_indices.R
\name{load_climate_index}
\alias{load_climate_index}
\title{Obtain data on large-scale climate oscillation indices.}
\usage{
load_climate_index(index)
}
\arguments{
\item{index}{One or more climate indices.
Supported indices include.
\itemize{
\item Dipole Mode index ("dmi")
\item Multivariate ENSO Index ("mei")
\item Oceanic Niño Index ("oni")
\item Southern Oscillation Index ("soi")
\item Pacific Decadal Oscillation ("pdo")
\item Atlantic Multidecadal Oscillation ("amo")
\item North Atlantic Oscillation ("nao")
\item Southern Annular Mode ("sam")
}}
}
\description{
Returns a named list of dplyr::tbl_df objects.
}
\examples{
indices <- load_climate_index(c("mei", "oni"))
}

