% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital-rates.R
\name{make_survivorship_trials}
\alias{make_survivorship_trials}
\title{Converts the stage-specific survivorship data to binomial (trials & successes).}
\usage{
make_survivorship_trials(df)
}
\arguments{
\item{df}{Name of the tbl_df containing survivorship data generated by the function stage_specific_survival.}
}
\description{
Converts the stage-specific survivorship data to binomial (trials & successes).
}
\examples{
surv_trials <- make_survivorship_trials(sss)
}

