% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vital_rates.R
\name{median_age_first_rep}
\alias{median_age_first_rep}
\title{Calculate median age at first reproduction}
\usage{
median_age_first_rep(b)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}
}
\value{
Output is a tbl_df that gives the median age at first reproduction for each study species in days and years. Sample size is also included.
}
\description{
Calculate median age at first reproduction
}
\examples{
mafr <- median_age_first_rep(lh)
}

