% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plhdbR.R
\docType{package}
\name{plhdbR}
\alias{plhdbR}
\alias{plhdbR-package}
\title{plhdbR aims to facillitate the analysis of data in the PLHDB.}
\description{
It will have three main sets of utilities:
}
\details{
\itemize{
\item Functions for reading and error-checking the life history and fertility data.
\item Functions for calculating vital rates.
\item Functions for loading and analyzing climate data.
}

Additional functionality may be added in the future.
}

