% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital-rates.R
\name{pseudocensus_ages}
\alias{pseudocensus_ages}
\title{Create pseudocensus dates and calculate ages of all animals present at each census.
This function is used by the stage-specific vital rate functions.}
\usage{
pseudocensus_ages(b, adult_definition = "median", census_start_month = NULL)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}

\item{adult_definition}{How adult females are defined in terms of age at first birth. Options are "median" or "minimum".}

\item{census_start_month}{A vector of start months for each study. Should be same length as the number of studies. Defaults to January.}
}
\value{
Output is a tbl_df that includes an entry for each animal present at each pseudo-census date along with its discrete age class and life-history stage at that date.
}
\description{
Create pseudocensus dates and calculate ages of all animals present at each census.
This function is used by the stage-specific vital rate functions.
}
\examples{
ages <- pseudocensus_ages(lh)
}

