% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_tables.R
\name{read_bio_table}
\alias{read_bio_table}
\title{Read biography table from plhdb.org csv export.}
\usage{
read_bio_table(f)
}
\arguments{
\item{f}{Name of the csv file containing biography data. Do not edit the file manually after downloading.}
}
\description{
Returns a tbl_df.
}
\examples{
# Assuming your file is called "biography_2015_03_17.csv"
lh <- read_bio_table("biography_2015_03_17.csv")
}

