% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tables.R
\name{read_fert_table}
\alias{read_fert_table}
\title{Read fertility table from plhdb.org csv export.}
\usage{
read_fert_table(f)
}
\arguments{
\item{f}{Name of the csv file containing fertility data. Do not edit the file manually after downloading.}
}
\description{
Returns a tbl_df.
}
\examples{
# Assuming your file is called "fertility_2015_03_17.csv"
lh <- read_fert_table("fertility_2015_03_17.csv")
}

