% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{sqrt_sign_trans}
\alias{sqrt_sign_trans}
\title{A signed square-root transformation for scales, e.g., color gradients.}
\usage{
sqrt_sign_trans()
}
\description{
A signed square-root transformation for scales, e.g., color gradients.
}

