% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital-rates.R
\name{stage_specific_fertility}
\alias{stage_specific_fertility}
\title{Calculate stage-specific fertility
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.}
\usage{
stage_specific_fertility(b, f, adult_definition = "median", annual = TRUE,
  census_start_month = NULL, filter_fertile = FALSE)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}

\item{f}{Name of the tbl_df containing fertility data generated by the function read_fert_table.}

\item{adult_definition}{Character. Either "median" or "minimum". Determines how females should be defined: minimum age at first birth or median age at first birth.}

\item{annual}{Logical. Should the fertility be calculated for each year separately?}

\item{census_start_month}{Numeric vector of length equal to number of studies.}

\item{filter_fertile}{Logical. If set to true, only females that are likely to be fertile are retained.}
}
\value{
Output is a tbl_df that includes the per capita fertility for each life-history stage for each study species. Life history stages include newborns, juveniles, and adults; these are defined as in Morris et al. 2011. Also included are the number of animals and the "female-years" of observation on which the calculation is based.
}
\description{
Calculate stage-specific fertility
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.
}
\examples{
ssf <- stage_specific_fertility(lh, fert)
}

