% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vital_rates.R
\name{stage_specific_mortality}
\alias{stage_specific_mortality}
\title{Calculate stage-specific mortality
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.}
\usage{
stage_specific_mortality(b)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}
}
\value{
Output is a tbl_df that includes the probability of survival for each life-history stage for each study species. Life history stages include newborns, juveniles, and adults; these are defined as in Morris et al. 2011. Also included are the number of animals and the "animal-years" of observation on which the calculation is based.
}
\description{
Calculate stage-specific mortality
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.
}
\examples{
ssm <- stage_specific_mortality(lh)
}

