% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vital_rates.R
\name{age_specific_fertility}
\alias{age_specific_fertility}
\title{Calculate age-specific fertility}
\usage{
age_specific_fertility(b, f)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}

\item{f}{Name of the tbl_df containing fertility data generated by the function read_fert_table.}
}
\description{
Returns a dplyr::tbl_df.
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.
}
\examples{
asf <- age_specific_fertility(lh, fert)
}

