% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vital_rates.R
\name{median_age_first_rep}
\alias{median_age_first_rep}
\title{Calculate median age at first reproduction}
\usage{
median_age_first_rep(b, f)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}

\item{f}{Name of the tbl_df containing fertility data generated by the function read_fert_table.}
}
\description{
Returns a dplyr::tbl_df.
}
\examples{
a <- median_age_first_rep(lh, fert)
}

