% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vital_rates.R
\name{stage_specific_fertility}
\alias{stage_specific_fertility}
\title{Calculate stage-specific fertility
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.}
\usage{
stage_specific_fertility(b, f, annual = TRUE)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}

\item{f}{Name of the tbl_df containing fertility data generated by the function read_fert_table.}

\item{f}{Logical. Should the fertility be calculated for each year separately?}
}
\description{
Returns a dplyr::tbl_df.
}
\examples{
ssf <- stage_specific_fertility(lh, fert)
}

