% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vital_rates.R
\name{stage_specific_fertility}
\alias{stage_specific_fertility}
\title{Calculate stage-specific fertility
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.}
\usage{
stage_specific_fertility(b, f, annual = TRUE)
}
\arguments{
\item{b}{Name of the tbl_df containing biography data generated by the function read_bio_table.}

\item{f}{Name of the tbl_df containing fertility data generated by the function read_fert_table.}

\item{f}{Logical. Should the fertility be calculated for each year separately?}
}
\value{
Output is a tbl_df that includes the per capita fertility for each life-history stage for each study species. Life history stages include newborns, juveniles, and adults; these are defined as in Morris et al. 2011. Also included are the number of animals and the "female-years" of observation on which the calculation is based.
}
\description{
Calculate stage-specific fertility
Based on Bill Morris' methods for calculating vital rates, available on the Wiki.
}
\examples{
ssf <- stage_specific_fertility(lh, fert)
}

